Refactor : 팔로우버튼 컴포넌트화
